Cleanup: comments, naming. Move methods.
